AP_HAL_ChibiOS: Increase limit for Storage pages
STm32l4+ has 512 flash pages in dual flash mode

+++ libraries/AP_HAL_ChibiOS/Storage.h
@@ -79,7 +79,7 @@ private:
     bool _flash_read_data(uint8_t sector, uint32_t offset, uint8_t *data, uint16_t length);
     bool _flash_erase_sector(uint8_t sector);
     bool _flash_erase_ok(void);
-    uint8_t _flash_page;
+    uint16_t _flash_page;
     bool _flash_failed;
     uint32_t _last_re_init_ms;
     uint32_t _last_empty_ms;
